+++ src/main/java/com/bunq/sdk/exception/ApiException.java
@@ -1,9 +1,34 @@
 package com.bunq.sdk.exception;
 
+public class ApiException extends RuntimeException {
 
-public class ApiException extends BunqError {
+  /**
+   */
+  private int responseCode;
+  private String message;
 
+  /**
+   */
+  public ApiException(){super();}
+
+  /**
+   * @param message The error message of the failed request.
+   * @param responseCode The response code of the failed request.
+   */
   public ApiException(String message, int responseCode) {
-    super(message, responseCode);
+    super(message);
+
+    this.responseCode = responseCode;
+    this.message = message;
   }
+
+  public int getResponseCode() {
+    return responseCode;
+  }
+
+  @Override
+  public String getMessage() {
+    return message;
+  }
+
 }
