+++ src/main/java/com/bunq/sdk/exception/BadRequestException.java
@@ -1,8 +1,9 @@
 package com.bunq.sdk.exception;
 
-public class BadRequestException extends BunqError {
+public class BadRequestException extends ApiException {
 
   public BadRequestException(String message, int responseCode) {
     super(message, responseCode);
   }
+
 }
