--- src/main/java/com/bunq/sdk/exception/BunqError.java
@@ -1,30 +1,0 @@
-package com.bunq.sdk.exception;
-
-public class BunqError extends RuntimeException {
-  private int responseCode;
-  private String message;
-
-  /**
-   */
-  public BunqError(){super();}
-
-  /**
-   * @param message The error message of the failed request.
-   * @param responseCode The response code of the failed request.
-   */
-  public BunqError(String message, int responseCode) {
-    super(message);
-
-    this.responseCode = responseCode;
-    this.message = message;
-  }
-
-  public int getResponseCode() {
-    return responseCode;
-  }
-
-  @Override
-  public String getMessage() {
-    return message;
-  }
-}
