+++ src/main/java/com/bunq/sdk/exception/ForbiddenException.java
@@ -1,8 +1,9 @@
 package com.bunq.sdk.exception;
 
-public class ForbiddenException extends BunqError {
+public class ForbiddenException extends ApiException {
 
   public ForbiddenException(String message, int responseCode) {
     super(message, responseCode);
   }
+
 }
