+++ src/main/java/com/bunq/sdk/exception/MethodNotAllowedException.java
@@ -1,8 +1,9 @@
 package com.bunq.sdk.exception;
 
-public class MethodNotAllowedException extends BunqError {
+public class MethodNotAllowedException extends ApiException {
 
   public MethodNotAllowedException(String message, int responseCode) {
     super(message, responseCode);
   }
+
 }
