+++ src/main/java/com/bunq/sdk/exception/NotFoundException.java
@@ -1,8 +1,9 @@
 package com.bunq.sdk.exception;
 
-public class NotFoundException extends BunqError {
+public class NotFoundException extends ApiException {
 
   public NotFoundException(String message, int responseCode) {
     super(message, responseCode);
   }
+
 }
