+++ src/main/java/com/bunq/sdk/exception/PleaseContactBunqException.java
@@ -1,8 +1,9 @@
 package com.bunq.sdk.exception;
 
-public class PleaseContactBunqException extends BunqError {
+public class PleaseContactBunqException extends ApiException {
 
   public PleaseContactBunqException(String message, int responseCode) {
     super(message, responseCode);
   }
+
 }
