+++ src/main/java/com/bunq/sdk/exception/ToManyRequestsException.java
@@ -1,8 +1,9 @@
 package com.bunq.sdk.exception;
 
-public class ToManyRequestsException extends BunqError {
+public class ToManyRequestsException extends ApiException {
 
   public ToManyRequestsException(String message, int responseCode) {
     super(message, responseCode);
   }
+
 }
