+++ src/main/java/com/bunq/sdk/exception/UnauthorizedException.java
@@ -1,8 +1,9 @@
 package com.bunq.sdk.exception;
 
-public class UnauthorizedException extends BunqError {
+public class UnauthorizedException extends ApiException {
 
   public UnauthorizedException(String message, int responseCode) {
     super(message, responseCode);
   }
+
 }
