+++ src/main/java/com/bunq/sdk/exception/UnknownApiErrorException.java
@@ -1,0 +1,10 @@
+package com.bunq.sdk.exception;
+
+
+public class UnknownApiErrorException extends ApiException {
+
+  public UnknownApiErrorException(String message, int responseCode) {
+    super(message, responseCode);
+  }
+
+}
